Update ex059_melhorado.py
coloquei número sendo float

+++ Exercicios/Ex059/ex059_melhorado.py
@@ -1,5 +1,5 @@
-n1 = int(input('Primeiro valor: '))
-n2 = int(input('Segundo valor: '))
+n1 = float(input('Primeiro valor: '))
+n2 = float(input('Segundo valor: '))
 opção = 0
 while opção != 5:
     print('''    [1] Somar
@@ -13,8 +13,8 @@ while opção != 5:
         print('A soma entre {} + {} é igual a {}'.format(n1, n2, soma))
         continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
         if continuar == 'S':
-            n1 = int(input('Primeiro valor: '))
-            n2 = int(input('Segundo valor: '))
+            n1 = float(input('Primeiro valor: '))
+            n2 = float(input('Segundo valor: '))
         elif continuar == 'N':
             continuar = 5
             opção = continuar
@@ -22,8 +22,8 @@ while opção != 5:
             print('\033[1;31mOpção inválida, por favor informar a opção novamente\033[m')
             continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
             if continuar == 'S':
-                n1 = int(input('Primeiro valor: '))
-                n2 = int(input('Segundo valor: '))
+                n1 = float(input('Primeiro valor: '))
+                n2 = float(input('Segundo valor: '))
             elif continuar == 'N':
                 continuar = 5
                 opção = continuar
@@ -32,8 +32,8 @@ while opção != 5:
         print('O resultado da multiplicação entre {} x {} é igual a {}'. format(n1, n2, produto))
         continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
         if continuar == 'S':
-            n1 = int(input('Primeiro valor: '))
-            n2 = int(input('Segundo valor: '))
+            n1 = float(input('Primeiro valor: '))
+            n2 = float(input('Segundo valor: '))
         elif continuar == 'N':
             continuar = 5
             opção = continuar
@@ -41,8 +41,8 @@ while opção != 5:
             print('\033[1;31mOpção inválida, por favor informar a opção novamente\033[m')
             continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
             if continuar == 'S':
-                n1 = int(input('Primeiro valor: '))
-                n2 = int(input('Segundo valor: '))
+                n1 = float(input('Primeiro valor: '))
+                n2 = float(input('Segundo valor: '))
             elif continuar == 'N':
                 continuar = 5
                 opção = continuar
@@ -54,8 +54,8 @@ while opção != 5:
         print('Entre {} e {} o maior valor é o {}'.format(n1, n2, maior))
         continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
         if continuar == 'S':
-            n1 = int(input('Primeiro valor: '))
-            n2 = int(input('Segundo valor: '))
+            n1 = float(input('Primeiro valor: '))
+            n2 = float(input('Segundo valor: '))
         elif continuar == 'N':
             continuar = 5
             opção = continuar
@@ -63,15 +63,15 @@ while opção != 5:
             print('\033[1;31mOpção inválida, por favor informar a opção novamente\033[m')
             continuar = str(input('Deseja fazer novamente [S/N]? ')).strip().upper()
             if continuar == 'S':
-                n1 = int(input('Primeiro valor: '))
-                n2 = int(input('Segundo valor: '))
+                n1 = float(input('Primeiro valor: '))
+                n2 = float(input('Segundo valor: '))
             elif continuar == 'N':
                 continuar = 5
                 opção = continuar
     elif opção == 4:
         print('Informe os valores novamente!')
-        n1 = int(input('Primeiro valor: '))
-        n2 = int(input('Segundo valor: '))
+        n1 = float(input('Primeiro valor: '))
+        n2 = float(input('Segundo valor: '))
     elif opção == 5:
         print('Finalizando...')
     else:
